separated components for template and ui
