--- src/components/WpIcons.vue
@@ -1,236 +1,0 @@
-<template>
-    <svg style="display: none;" xmlns="http://www.w3.org/2000/svg">
-        <symbol id="skIconsHamburger" viewBox="0 96 960 960">
-            <path d="M120 816v-60h720v60H120Zm0-210v-60h720v60H120Zm0-210v-60h720v60H120Z"></path>
-        </symbol>
-        <symbol id="skIconsCall" viewBox="0 0 24 24">
-            <path d="M0 0h24v24H0z" fill="none"></path>
-            <path
-                d="M20.01 15.38c-1.23 0-2.42-.2-3.53-.56-.35-.12-.74-.03-1.01.24l-1.57 1.97c-2.83-1.35-5.48-3.9-6.89-6.83l1.95-1.66c.27-.28.35-.67.24-1.02-.37-1.11-.56-2.3-.56-3.53 0-.54-.45-.99-.99-.99H4.19C3.65 3 3 3.24 3 3.99 3 13.28 10.73 21 20.01 21c.71 0 .99-.63.99-1.18v-3.45c0-.54-.45-.99-.99-.99z">
-            </path>
-        </symbol>
-        <symbol id="skIconsDropDown" viewBox="0 0 24 24">
-            <path d="M0 0h24v24H0z" fill="none"></path>
-            <path d="M7 10l5 5 5-5z"></path>
-        </symbol>
-        <symbol id="skIconsLocation" viewBox="0 0 24 24">
-            <path d="M0 0h24v24H0z" fill="none"></path>
-            <path
-                d="M12 2C8.13 2 5 5.13 5 9c0 5.25 7 13 7 13s7-7.75 7-13c0-3.87-3.13-7-7-7zm0 9.5c-1.38 0-2.5-1.12-2.5-2.5s1.12-2.5 2.5-2.5 2.5 1.12 2.5 2.5-1.12 2.5-2.5 2.5z">
-            </path>
-        </symbol>
-        <symbol id="skIconDownChevron" viewBox="0 0 48 48">
-            <path d="m24 30.75-12-12 2.15-2.15L24 26.5l9.85-9.85L36 18.8Z"></path>
-        </symbol>
-        <symbol id="skIconsGetDirection" viewBox="0 0 24 24">
-            <path fill="none" d="M0,0h24v24H0V0z"></path>
-            <path d="M22.5,11.2l-9.8-9.8c-0.5-0.4-1.1-0.4-1.6,0l-9.8,9.8c-0.4,0.5-0.4,1.1,0,1.6l9.8,9.8c0.5,0.4,1.1,0.4,1.6,0l9.8-9.8
-    C23,12.3,23,11.6,22.5,11.2z M14,15.4l-1.9-1.9l1.5-1.5H9.3v4.1H6.6v-5.4c0-0.8,0.5-1.4,1.4-1.4h5.7l-1.5-1.5L14,5.9l4.8,4.8
-    L14,15.4z"></path>
-        </symbol>
-        <symbol id="skIconsShare" viewBox="0 0 24 24">
-            <path d="M0 0h24v24H0z" fill="none"></path>
-            <path
-                d="M18 16.08c-.76 0-1.44.3-1.96.77L8.91 12.7c.05-.23.09-.46.09-.7s-.04-.47-.09-.7l7.05-4.11c.54.5 1.25.81 2.04.81 1.66 0 3-1.34 3-3s-1.34-3-3-3-3 1.34-3 3c0 .24.04.47.09.7L8.04 9.81C7.5 9.31 6.79 9 6 9c-1.66 0-3 1.34-3 3s1.34 3 3 3c.79 0 1.5-.31 2.04-.81l7.12 4.16c-.05.21-.08.43-.08.65 0 1.61 1.31 2.92 2.92 2.92 1.61 0 2.92-1.31 2.92-2.92s-1.31-2.92-2.92-2.92z">
-            </path>
-        </symbol>
-        <symbol id="skIconsWatch" viewBox="0 0 24 24">
-            <rect fill="none" height="24" width="24"></rect>
-            <path
-                d="M12,2C6.5,2,2,6.5,2,12s4.5,10,10,10s10-4.5,10-10S17.5,2,12,2z M16.2,16.2L11,13V7h1.5v5.2l4.5,2.7L16.2,16.2z">
-            </path>
-        </symbol>
-        <symbol id="skIconsLocation" viewBox="0 0 24 24">
-            <path d="M0 0h24v24H0z" fill="none"></path>
-            <path
-                d="M12 2C8.13 2 5 5.13 5 9c0 5.25 7 13 7 13s7-7.75 7-13c0-3.87-3.13-7-7-7zm0 9.5c-1.38 0-2.5-1.12-2.5-2.5s1.12-2.5 2.5-2.5 2.5 1.12 2.5 2.5-1.12 2.5-2.5 2.5z">
-            </path>
-        </symbol>
-        <symbol id="skIconsNavPrev" viewBox="0 0 24 24">
-            <path fill="none" d="M0,0h24v24H0V0z"></path>
-            <path
-                d="M15.3,6.7c0.4,0.4,0.4,1,0,1.4L11.4,12l3.9,3.9c0.4,0.4,0.4,1,0,1.4c-0.4,0.4-1,0.4-1.4,0l-4.6-4.6c-0.4-0.4-0.4-1,0-1.4l4.6-4.6C14.3,6.3,14.9,6.3,15.3,6.7z">
-            </path>
-        </symbol>
-
-        <symbol id="skIconsNavNext" viewBox="0 0 24 24">
-            <path d="M0 0h24v24H0V0z" fill="none"></path>
-            <path
-                d="M9.31 6.71c-.39.39-.39 1.02 0 1.41L13.19 12l-3.88 3.88c-.39.39-.39 1.02 0 1.41.39.39 1.02.39 1.41 0l4.59-4.59c.39-.39.39-1.02 0-1.41L10.72 6.7c-.38-.38-1.02-.38-1.41.01z">
-            </path>
-        </symbol>
-        <symbol id="skIconsClock" viewBox="0 0 25 25">
-            <path fill-rule="evenodd" clip-rule="evenodd"
-                d="M0 12.6283C0 5.95481 5.40603 0.544922 12.0747 0.544922C18.7433 0.544922 24.1494 5.95481 24.1494 12.6283C24.1494 19.3016 18.7433 24.7116 12.0747 24.7116C5.40603 24.7116 0 19.3016 0 12.6283ZM16.7019 15.7152C16.3319 16.2704 15.5822 16.4205 15.0273 16.0504L11.8877 13.9557C11.2501 13.5304 10.8672 12.8145 10.8672 12.0478V12.0241V6.58659C10.8672 5.91925 11.4078 5.37826 12.0747 5.37826C12.7416 5.37826 13.2822 5.91925 13.2822 6.58659V11.9816L16.3669 14.0395C16.9217 14.4097 17.0717 15.16 16.7019 15.7152Z"
-                fill=""></path>
-        </symbol>
-
-        <!-- payment options icons -->
-        <symbol id="skIconsMasterCard" viewBox="0 0 40 40">
-            <g id="Group_2115" data-name="Group 2115" transform="translate(-960 -3398)">
-                <rect id="Rectangle_359" data-name="Rectangle 359" width="40" height="40"
-                    transform="translate(960 3398)" fill="none"></rect>
-                <g id="_728973_credit_card_mastercard_payment_card_pay_icon"
-                    data-name="728973_credit card_mastercard_payment_card_pay_icon" transform="translate(964 3390)">
-                    <g id="Group_2033" data-name="Group 2033" transform="translate(0 16)">
-                        <g id="Group_2032" data-name="Group 2032">
-                            <path id="Path_15952" data-name="Path 15952"
-                                d="M28,16H4a4,4,0,0,0-4,4V36a4,4,0,0,0,4,4H28a4,4,0,0,0,4-4V20A4,4,0,0,0,28,16Zm2,20a2,2,0,0,1-2,2H4a2,2,0,0,1-2-2V20a2,2,0,0,1,2-2H28a2,2,0,0,1,2,2Z"
-                                transform="translate(0 -16)"></path>
-                        </g>
-                    </g>
-                    <path id="Path_15953" data-name="Path 15953"
-                        d="M65.178,44.589A4.589,4.589,0,1,1,60.589,40,4.589,4.589,0,0,1,65.178,44.589Z"
-                        transform="translate(-41.637 -16.398)" fill="#ffa000"></path>
-                    <path id="Path_15954" data-name="Path 15954"
-                        d="M33.559,44.779A4.779,4.779,0,1,1,28.779,40,4.78,4.78,0,0,1,33.559,44.779Z"
-                        transform="translate(-16.559 -16.779)" fill="#d32f2f"></path>
-                </g>
-            </g>
-        </symbol>
-        <symbol id="skIconsOnlinePayment" viewBox="0 0 40 40">
-            <g id="Group_2114" data-name="Group 2114" transform="translate(-960 -3455)">
-                <rect id="Rectangle_360" data-name="Rectangle 360" width="40" height="40"
-                    transform="translate(960 3455)" fill="none"></rect>
-                <g id="Layer_x0020_1" transform="translate(954.93 3436.8)">
-                    <path id="Path_15955" data-name="Path 15955"
-                        d="M32.775,53.439a3.463,3.463,0,0,1-1.307.263H12.512A3.45,3.45,0,0,1,9.07,50.259V26.642A3.45,3.45,0,0,1,12.512,23.2H31.468a3.45,3.45,0,0,1,3.442,3.442V42.873c.174-.016.351-.025.53-.025a5.63,5.63,0,1,1-2.664,10.591Zm-.265-4.625a.8.8,0,0,1,1.295-.938L34.624,49l2.5-2.858a.8.8,0,0,1,1.2,1.051L35.2,50.77a.8.8,0,0,1-1.283-.016Zm.8-5.549V28.782H10.671V50.259A1.848,1.848,0,0,0,12.512,52.1H31.129a5.632,5.632,0,0,1,2.179-8.835Zm2.131,1.184a4.029,4.029,0,1,0,4.029,4.029A4.029,4.029,0,0,0,35.44,44.449ZM16.78,49.7a.8.8,0,0,1,0-1.6H27.2a.8.8,0,0,1,0,1.6ZM33.309,27.181v-.539A1.848,1.848,0,0,0,31.468,24.8H12.512a1.848,1.848,0,0,0-1.841,1.841v.539H33.309Z">
-                    </path>
-                </g>
-            </g>
-        </symbol>
-        <symbol id="skIconsVisa" viewBox="0 0 40 40">
-            <g id="Group_2113" data-name="Group 2113" transform="translate(-960 -3518)">
-                <rect id="Rectangle_361" data-name="Rectangle 361" width="40" height="40"
-                    transform="translate(960 3518)" fill="none"></rect>
-                <g id="visa-svgrepo-com" transform="translate(962 3521)">
-                    <path id="Path_16029" data-name="Path 16029"
-                        d="M30.571,28.857H5.429A3.432,3.432,0,0,1,2,25.429v-16A3.432,3.432,0,0,1,5.429,6H30.571A3.432,3.432,0,0,1,34,9.429v16A3.432,3.432,0,0,1,30.571,28.857ZM5.429,8.286A1.144,1.144,0,0,0,4.286,9.429v16a1.144,1.144,0,0,0,1.143,1.143H30.571a1.144,1.144,0,0,0,1.143-1.143v-16a1.144,1.144,0,0,0-1.143-1.143Z"
-                        transform="translate(0 0)"></path>
-                    <g id="Group_2112" data-name="Group 2112" transform="translate(5.238 13.281)">
-                        <path id="Path_16030" data-name="Path 16030"
-                            d="M9.923,16.883,9.218,13.3a1.3,1.3,0,0,0-1.146-.777H4.833S8.962,13.787,9.923,16.883Z"
-                            transform="translate(-4.833 -12.349)"></path>
-                        <path id="Path_16031" data-name="Path 16031" d="M12.782,20.4h2.094l1.266-7.89H14.023Z"
-                            transform="translate(-3.697 -12.351)"></path>
-                        <path id="Path_16032" data-name="Path 16032"
-                            d="M11.8,12.51,9.885,18.129s-.11-.479-.157-.776a7.056,7.056,0,0,0-2.862-3.342L8.548,20.4h2.306l3.186-7.889Z"
-                            transform="translate(-4.542 -12.351)"></path>
-                        <path id="Path_16033" data-name="Path 16033"
-                            d="M18.69,14.789c0-.457.363-.8,1.4-.8a3.229,3.229,0,0,1,1.453.51l.339-1.743a6.237,6.237,0,0,0-1.962-.387c-2.2,0-3.338,1.089-3.338,2.467,0,2.493,2.9,2.152,2.9,3.433,0,.218-.169.726-1.378.726a4.348,4.348,0,0,1-2.01-.461l-.362,1.672a4.78,4.78,0,0,0,2.275.458c1.5,0,3.582-1.161,3.582-2.831C21.6,15.828,18.69,15.684,18.69,14.789Z"
-                            transform="translate(-3.275 -12.371)"></path>
-                        <path id="Path_16034" data-name="Path 16034"
-                            d="M26.453,12.51h-2.2L20.814,20.4h2.079l.431-1.126h2.621l.225,1.126h1.905ZM23.9,17.766l1.138-2.981.6,2.981Z"
-                            transform="translate(-2.55 -12.351)"></path>
-                    </g>
-                </g>
-            </g>
-        </symbol>
-        <symbol id="skIconsDebitCard" viewBox="0 0 40 40">
-            <g id="Group_2111" data-name="Group 2111" transform="translate(-960 -3337)">
-                <rect id="Rectangle_358" data-name="Rectangle 358" width="40" height="40"
-                    transform="translate(960 3337)" fill="none"></rect>
-                <path id="_7076551_pay_debit_card_payment_card_credit_card_icon"
-                    data-name="7076551_pay_debit card_payment_card_credit card_icon"
-                    d="M30.474,11.5H3.014a2.273,2.273,0,0,0-2.27,2.27V30.217a2.273,2.273,0,0,0,2.27,2.271h27.46a2.273,2.273,0,0,0,2.27-2.271V13.774A2.273,2.273,0,0,0,30.474,11.5ZM3.014,13.04h27.46a.735.735,0,0,1,.734.735v1.471H2.28V13.774A.735.735,0,0,1,3.014,13.04Zm28.194,5.885H2.28V16.781H31.208Zm-.734,12.027H3.014a.736.736,0,0,1-.735-.735V20.46H31.208v9.757A.735.735,0,0,1,30.474,30.952Z"
-                    transform="translate(963.256 3335.496)"></path>
-            </g>
-        </symbol>
-        <symbol id="skIconsCheque" viewBox="0 0 40 40">
-            <g id="Group_2110" data-name="Group 2110" transform="translate(-960 -3279)">
-                <rect id="Rectangle_357" data-name="Rectangle 357" width="40" height="40"
-                    transform="translate(960 3279)" fill="none"></rect>
-                <g id="_36._Bank_Check" data-name="36. Bank Check" transform="translate(964 3283)">
-                    <path id="Path_15948" data-name="Path 15948"
-                        d="M31,8H28V3a3,3,0,1,0-6,0V8H3a3,3,0,0,0-3,3V29a3,3,0,0,0,3,3H23a1,1,0,0,0,1-1V28h7a1,1,0,0,0,1-1V9A1,1,0,0,0,31,8ZM24,6h2v8.59l-1,1-1-1Zm1-4a1,1,0,0,1,1,1V4H24V3A1,1,0,0,1,25,2ZM22,30H3a1,1,0,0,1,0-2H22Zm8-4H3a3,3,0,0,0-1,.18V11a1,1,0,0,1,1-1H22v5a1,1,0,0,0,.29.71l2,2a1,1,0,0,0,1.42,0l2-2A1,1,0,0,0,28,15V10h2Z">
-                    </path>
-                    <path id="Path_15949" data-name="Path 15949" d="M5,14h8a1,1,0,0,0,0-2H5a1,1,0,0,0,0,2Z"></path>
-                    <path id="Path_15950" data-name="Path 15950"
-                        d="M18,17a1,1,0,0,0-1-1H5a1,1,0,0,0,0,2H17A1,1,0,0,0,18,17Z"></path>
-                    <path id="Path_15951" data-name="Path 15951"
-                        d="M27,22H25.54l-1.71-2.55a1,1,0,0,0-1.66,0L20.46,22H18a1,1,0,0,0,0,2h3a1,1,0,0,0,.83-.45L23,21.8l1.17,1.75A1,1,0,0,0,25,24h2a1,1,0,0,0,0-2Z">
-                    </path>
-                </g>
-            </g>
-        </symbol>
-        <symbol id="skIconsCash" viewBox="0 0 40 40">
-            <g id="Group_2109" data-name="Group 2109" transform="translate(-960 -3217)">
-                <rect id="Rectangle_356" data-name="Rectangle 356" width="40" height="40"
-                    transform="translate(960 3217)" fill="none"></rect>
-                <path id="_290143_cash_money_payment_wallet_icon" data-name="290143_cash_money_payment_wallet_icon"
-                    d="M34.578,30.28h0a3.649,3.649,0,0,1-3.65,3.65H4.65A3.649,3.649,0,0,1,1,30.28V9.111a2.92,2.92,0,0,1,2.92-2.92H18.644L24.556,4.05a.723.723,0,0,1,.926.43L26.1,6.192h2.634a2.92,2.92,0,0,1,2.92,2.92v1.46h0a2.92,2.92,0,0,1,2.92,2.92ZM3.92,7.651h0a1.46,1.46,0,1,0,0,2.92H6.573c.009,0,.015-.012.025-.015l8.016-2.9Zm21.226.132L24.37,5.649h0l-1.5.542h0L18.85,7.651h-.007l-8.058,2.92H26.159ZM30.2,9.111a1.46,1.46,0,0,0-1.46-1.46h-2.1l1.061,2.92h2.5Zm1.46,2.92H3.92a2.887,2.887,0,0,1-1.46-.405V30.28a2.19,2.19,0,0,0,2.19,2.19H30.928a2.19,2.19,0,0,0,2.19-2.19V25.17H30.2a2.92,2.92,0,1,1,0-5.84h2.92v-5.84A1.46,1.46,0,0,0,31.658,12.031Zm1.46,11.679v-2.92H30.2a1.46,1.46,0,1,0,0,2.92Zm-2.92-2.19h1.46v1.46H30.2Z"
-                    transform="translate(962 3216.994)" fill-rule="evenodd"></path>
-            </g>
-        </symbol>
-
-        <!-- payment options icons -->
-
-        <symbol id="skIconsCalendar" viewBox="0 -960 960 960">
-            <path
-                d="M200-80q-33 0-56.5-23.5T120-160v-560q0-33 23.5-56.5T200-800h40v-80h80v80h320v-80h80v80h40q33 0 56.5 23.5T840-720v560q0 33-23.5 56.5T760-80H200Zm0-80h560v-400H200v400Zm0-480h560v-80H200v80Zm0 0v-80 80Zm280 240q-17 0-28.5-11.5T440-440q0-17 11.5-28.5T480-480q17 0 28.5 11.5T520-440q0 17-11.5 28.5T480-400Zm-160 0q-17 0-28.5-11.5T280-440q0-17 11.5-28.5T320-480q17 0 28.5 11.5T360-440q0 17-11.5 28.5T320-400Zm320 0q-17 0-28.5-11.5T600-440q0-17 11.5-28.5T640-480q17 0 28.5 11.5T680-440q0 17-11.5 28.5T640-400ZM480-240q-17 0-28.5-11.5T440-280q0-17 11.5-28.5T480-320q17 0 28.5 11.5T520-280q0 17-11.5 28.5T480-240Zm-160 0q-17 0-28.5-11.5T280-280q0-17 11.5-28.5T320-320q17 0 28.5 11.5T360-280q0 17-11.5 28.5T320-240Zm320 0q-17 0-28.5-11.5T600-280q0-17 11.5-28.5T640-320q17 0 28.5 11.5T680-280q0 17-11.5 28.5T640-240Z" />
-        </symbol>
-        <symbol id="skiconsBranch" viewBox="0 0 19 17">
-            <path
-                d="M1.16667 2.24999V0.166656H17.8333V2.24999H1.16667ZM1.16667 16.8333V10.5833H0.125V8.49999L1.16667 3.29166H17.8333L18.875 8.49999V10.5833H17.8333V16.8333H15.75V10.5833H11.5833V16.8333H1.16667ZM3.25 14.75H9.5V10.5833H3.25V14.75ZM2.26042 8.49999H16.7396L16.1146 5.37499H2.88542L2.26042 8.49999Z" />
-        </symbol>
-
-        <symbol id="skiconsAlaram" viewBox="0 0 20 20">
-            <g clip-path="url(#clip0_472_140)">
-                <path
-                    d="M8.16671 13.5L6.50004 11.875C6.35568 11.7275 6.26813 11.5336 6.25288 11.3277C6.23763 11.1219 6.29566 10.9172 6.41671 10.75C6.48415 10.6496 6.57371 10.566 6.67851 10.5055C6.7833 10.4451 6.90055 10.4095 7.02125 10.4015C7.14195 10.3934 7.26289 10.4132 7.37478 10.4591C7.48667 10.5051 7.58653 10.5761 7.66671 10.6667L8.75004 11.75L12.3334 8.16668C12.4939 8.04633 12.6924 7.98789 12.8924 8.00211C13.0925 8.01633 13.2808 8.10225 13.4226 8.24409C13.5645 8.38594 13.6504 8.57418 13.6646 8.77428C13.6788 8.97437 13.6204 9.17287 13.5 9.33335L9.33338 13.5C9.25922 13.5808 9.16907 13.6454 9.06865 13.6895C8.96823 13.7337 8.85974 13.7565 8.75004 13.7565C8.64035 13.7565 8.53185 13.7337 8.43143 13.6895C8.33102 13.6454 8.24087 13.5808 8.16671 13.5Z"
-                    fill="black" />
-                <path
-                    d="M9.99968 5.00001C11.2358 5.00001 12.4442 5.36657 13.472 6.05332C14.4998 6.74008 15.3009 7.7162 15.7739 8.85824C16.247 10.0003 16.3707 11.2569 16.1296 12.4693C15.8884 13.6817 15.2932 14.7953 14.4191 15.6694C13.545 16.5435 12.4314 17.1388 11.219 17.3799C10.0066 17.6211 8.74994 17.4973 7.60791 17.0243C6.46587 16.5512 5.48975 15.7501 4.80299 14.7223C4.11623 13.6945 3.74968 12.4861 3.74968 11.25C3.74968 9.59241 4.40816 8.00269 5.58026 6.83059C6.75236 5.65849 8.34207 5.00001 9.99968 5.00001ZM9.99968 3.33334C8.43391 3.33334 6.9033 3.79765 5.60141 4.66754C4.29952 5.53744 3.28483 6.77385 2.68563 8.22043C2.08644 9.66701 1.92966 11.2588 2.23513 12.7945C2.54059 14.3302 3.29458 15.7408 4.40175 16.8479C5.50891 17.9551 6.91953 18.7091 8.45521 19.0146C9.9909 19.32 11.5827 19.1632 13.0293 18.5641C14.4758 17.9649 15.7123 16.9502 16.5821 15.6483C17.452 14.3464 17.9163 12.8158 17.9163 11.25C17.9163 9.15038 17.0823 7.13674 15.5976 5.65208C14.1129 4.16742 12.0993 3.33334 9.99968 3.33334Z"
-                    fill="black" />
-                <path
-                    d="M2.49984 5.00013C2.37103 4.99708 2.24448 4.96562 2.12924 4.908C2.01399 4.85038 1.9129 4.76802 1.83317 4.6668C1.76751 4.57925 1.71974 4.47963 1.69258 4.37362C1.66542 4.26761 1.6594 4.15728 1.67488 4.04895C1.70614 3.83016 1.82303 3.63274 1.99984 3.50013L5.33317 1.00013C5.42072 0.934471 5.52034 0.886697 5.62635 0.859538C5.73236 0.832378 5.84269 0.826365 5.95102 0.841841C6.05936 0.857317 6.16358 0.89398 6.25775 0.949737C6.35191 1.00549 6.43418 1.07925 6.49984 1.1668C6.5655 1.25435 6.61327 1.35397 6.64043 1.45998C6.66759 1.56599 6.6736 1.67632 6.65813 1.78465C6.64265 1.89299 6.60599 1.99721 6.55023 2.09138C6.49448 2.18554 6.42072 2.2678 6.33317 2.33347L2.99984 4.83347C2.8571 4.9445 2.68065 5.00331 2.49984 5.00013Z"
-                    fill="black" />
-                <path
-                    d="M17.4997 5.00001C17.3189 5.0032 17.1424 4.94438 16.9997 4.83335L13.6663 2.33335C13.5788 2.26769 13.505 2.18542 13.4493 2.09126C13.3935 1.99709 13.3569 1.89287 13.3414 1.78453C13.3259 1.6762 13.3319 1.56587 13.3591 1.45986C13.3862 1.35385 13.434 1.25423 13.4997 1.16668C13.6323 0.989869 13.8297 0.872978 14.0485 0.841722C14.2673 0.810466 14.4895 0.867405 14.6663 1.00001L17.9997 3.50001C18.1765 3.63262 18.2934 3.83004 18.3246 4.04883C18.3559 4.26762 18.2989 4.48987 18.1663 4.66668C18.0866 4.7679 17.9855 4.85026 17.8703 4.90788C17.755 4.9655 17.6285 4.99696 17.4997 5.00001Z"
-                    fill="black" />
-            </g>
-            <defs>
-                <clipPath id="clip0_472_140">
-                    <rect width="20" height="20" fill="white" />
-                </clipPath>
-            </defs>
-        </symbol>
-        <symbol id="skIconsLinkedin" viewBox="0 0 24 24">
-            <path
-                d="M18.72 3.99997H5.37C5.19793 3.99191 5.02595 4.01786 4.86392 4.07635C4.70189 4.13484 4.55299 4.22471 4.42573 4.34081C4.29848 4.45692 4.19537 4.59699 4.12232 4.75299C4.04927 4.909 4.0077 5.07788 4 5.24997V18.63C4.01008 18.9901 4.15766 19.3328 4.41243 19.5875C4.6672 19.8423 5.00984 19.9899 5.37 20H18.72C19.0701 19.9844 19.4002 19.8322 19.6395 19.5761C19.8788 19.32 20.0082 18.9804 20 18.63V5.24997C20.0029 5.08247 19.9715 4.91616 19.9078 4.76122C19.8441 4.60629 19.7494 4.466 19.6295 4.34895C19.5097 4.23191 19.3672 4.14059 19.2108 4.08058C19.0544 4.02057 18.8874 3.99314 18.72 3.99997ZM9 17.34H6.67V10.21H9V17.34ZM7.89 9.12997C7.72741 9.13564 7.5654 9.10762 7.41416 9.04768C7.26291 8.98774 7.12569 8.89717 7.01113 8.78166C6.89656 8.66615 6.80711 8.5282 6.74841 8.37647C6.6897 8.22474 6.66301 8.06251 6.67 7.89997C6.66281 7.73567 6.69004 7.57169 6.74995 7.41854C6.80986 7.26538 6.90112 7.12644 7.01787 7.01063C7.13463 6.89481 7.2743 6.80468 7.42793 6.74602C7.58157 6.68735 7.74577 6.66145 7.91 6.66997C8.07259 6.66431 8.2346 6.69232 8.38584 6.75226C8.53709 6.8122 8.67431 6.90277 8.78887 7.01828C8.90344 7.13379 8.99289 7.27174 9.05159 7.42347C9.1103 7.5752 9.13699 7.73743 9.13 7.89997C9.13719 8.06427 9.10996 8.22825 9.05005 8.3814C8.99014 8.53456 8.89888 8.6735 8.78213 8.78931C8.66537 8.90513 8.5257 8.99526 8.37207 9.05392C8.21843 9.11259 8.05423 9.13849 7.89 9.12997ZM17.34 17.34H15V13.44C15 12.51 14.67 11.87 13.84 11.87C13.5822 11.8722 13.3313 11.9541 13.1219 12.1045C12.9124 12.2549 12.7546 12.4664 12.67 12.71C12.605 12.8926 12.5778 13.0865 12.59 13.28V17.34H10.29V10.21H12.59V11.21C12.7945 10.8343 13.0988 10.5225 13.4694 10.3089C13.84 10.0954 14.2624 9.98848 14.69 9.99997C16.2 9.99997 17.34 11 17.34 13.13V17.34Z">
-            </path>
-        </symbol>
-        <symbol id="skIconsLinkedin" viewBox="0 0 24 24">
-            <path
-                d="M18.72 3.99997H5.37C5.19793 3.99191 5.02595 4.01786 4.86392 4.07635C4.70189 4.13484 4.55299 4.22471 4.42573 4.34081C4.29848 4.45692 4.19537 4.59699 4.12232 4.75299C4.04927 4.909 4.0077 5.07788 4 5.24997V18.63C4.01008 18.9901 4.15766 19.3328 4.41243 19.5875C4.6672 19.8423 5.00984 19.9899 5.37 20H18.72C19.0701 19.9844 19.4002 19.8322 19.6395 19.5761C19.8788 19.32 20.0082 18.9804 20 18.63V5.24997C20.0029 5.08247 19.9715 4.91616 19.9078 4.76122C19.8441 4.60629 19.7494 4.466 19.6295 4.34895C19.5097 4.23191 19.3672 4.14059 19.2108 4.08058C19.0544 4.02057 18.8874 3.99314 18.72 3.99997ZM9 17.34H6.67V10.21H9V17.34ZM7.89 9.12997C7.72741 9.13564 7.5654 9.10762 7.41416 9.04768C7.26291 8.98774 7.12569 8.89717 7.01113 8.78166C6.89656 8.66615 6.80711 8.5282 6.74841 8.37647C6.6897 8.22474 6.66301 8.06251 6.67 7.89997C6.66281 7.73567 6.69004 7.57169 6.74995 7.41854C6.80986 7.26538 6.90112 7.12644 7.01787 7.01063C7.13463 6.89481 7.2743 6.80468 7.42793 6.74602C7.58157 6.68735 7.74577 6.66145 7.91 6.66997C8.07259 6.66431 8.2346 6.69232 8.38584 6.75226C8.53709 6.8122 8.67431 6.90277 8.78887 7.01828C8.90344 7.13379 8.99289 7.27174 9.05159 7.42347C9.1103 7.5752 9.13699 7.73743 9.13 7.89997C9.13719 8.06427 9.10996 8.22825 9.05005 8.3814C8.99014 8.53456 8.89888 8.6735 8.78213 8.78931C8.66537 8.90513 8.5257 8.99526 8.37207 9.05392C8.21843 9.11259 8.05423 9.13849 7.89 9.12997ZM17.34 17.34H15V13.44C15 12.51 14.67 11.87 13.84 11.87C13.5822 11.8722 13.3313 11.9541 13.1219 12.1045C12.9124 12.2549 12.7546 12.4664 12.67 12.71C12.605 12.8926 12.5778 13.0865 12.59 13.28V17.34H10.29V10.21H12.59V11.21C12.7945 10.8343 13.0988 10.5225 13.4694 10.3089C13.84 10.0954 14.2624 9.98848 14.69 9.99997C16.2 9.99997 17.34 11 17.34 13.13V17.34Z">
-            </path>
-        </symbol>
-        <symbol id="skIconsInstagram" viewBox="0 0 24 24">
-            <path
-                d="M12 7.90001C11.1891 7.90001 10.3964 8.14048 9.72218 8.59099C9.04794 9.0415 8.52243 9.68184 8.21211 10.431C7.90179 11.1802 7.8206 12.0046 7.9788 12.7999C8.13699 13.5952 8.52748 14.3258 9.10088 14.8992C9.67427 15.4725 10.4048 15.863 11.2001 16.0212C11.9955 16.1794 12.8198 16.0982 13.569 15.7879C14.3182 15.4776 14.9585 14.9521 15.409 14.2779C15.8596 13.6036 16.1 12.8109 16.1 12C16.1013 11.4612 15.9962 10.9275 15.7906 10.4295C15.585 9.93142 15.2831 9.47892 14.9021 9.09794C14.5211 8.71695 14.0686 8.415 13.5706 8.20942C13.0725 8.00385 12.5388 7.8987 12 7.90001ZM12 14.67C11.4719 14.67 10.9557 14.5134 10.5166 14.22C10.0776 13.9267 9.73534 13.5097 9.53326 13.0218C9.33117 12.5339 9.2783 11.9971 9.38132 11.4791C9.48434 10.9612 9.73863 10.4854 10.112 10.112C10.4854 9.73863 10.9612 9.48434 11.4791 9.38132C11.9971 9.2783 12.5339 9.33117 13.0218 9.53326C13.5097 9.73534 13.9267 10.0776 14.22 10.5166C14.5134 10.9557 14.67 11.4719 14.67 12C14.67 12.7081 14.3887 13.3873 13.888 13.888C13.3873 14.3887 12.7081 14.67 12 14.67ZM17.23 7.73001C17.23 7.9278 17.1714 8.12114 17.0615 8.28558C16.9516 8.45003 16.7954 8.57821 16.6127 8.65389C16.43 8.72958 16.2289 8.74938 16.0349 8.7108C15.8409 8.67221 15.6628 8.57697 15.5229 8.43712C15.3831 8.29727 15.2878 8.11909 15.2492 7.92511C15.2106 7.73112 15.2304 7.53006 15.3061 7.34733C15.3818 7.16461 15.51 7.00843 15.6744 6.89855C15.8389 6.78866 16.0322 6.73001 16.23 6.73001C16.4952 6.73001 16.7496 6.83537 16.9371 7.02291C17.1247 7.21044 17.23 7.4648 17.23 7.73001ZM19.94 8.73001C19.9691 7.48684 19.5054 6.28261 18.65 5.38001C17.7522 4.5137 16.5474 4.03897 15.3 4.06001C14 4.00001 10 4.00001 8.70001 4.06001C7.45722 4.0331 6.25379 4.49652 5.35001 5.35001C4.49465 6.25261 4.03093 7.45684 4.06001 8.70001C4.00001 10 4.00001 14 4.06001 15.3C4.03093 16.5432 4.49465 17.7474 5.35001 18.65C6.25379 19.5035 7.45722 19.9669 8.70001 19.94C10.02 20.02 13.98 20.02 15.3 19.94C16.5432 19.9691 17.7474 19.5054 18.65 18.65C19.5054 17.7474 19.9691 16.5432 19.94 15.3C20 14 20 10 19.94 8.70001V8.73001ZM18.24 16.73C18.1042 17.074 17.8993 17.3863 17.6378 17.6478C17.3763 17.9093 17.064 18.1142 16.72 18.25C15.1676 18.5639 13.5806 18.6715 12 18.57C10.4228 18.6716 8.83902 18.564 7.29001 18.25C6.94608 18.1142 6.63369 17.9093 6.37223 17.6478C6.11076 17.3863 5.90579 17.074 5.77001 16.73C5.35001 15.67 5.44001 13.17 5.44001 12.01C5.44001 10.85 5.35001 8.34001 5.77001 7.29001C5.90196 6.94268 6.10547 6.62698 6.36733 6.36339C6.62919 6.09981 6.94355 5.89423 7.29001 5.76001C8.83902 5.44599 10.4228 5.33839 12 5.44001C13.5806 5.33856 15.1676 5.44616 16.72 5.76001C17.064 5.89579 17.3763 6.10076 17.6378 6.36223C17.8993 6.62369 18.1042 6.93608 18.24 7.28001C18.66 8.34001 18.56 10.84 18.56 12C18.56 13.16 18.66 15.67 18.24 16.72V16.73Z">
-            </path>
-        </symbol>
-        <symbol id="skIconsCalendarAdd" viewBox="0 -960 960 960">
-            <path
-                d="M680-80v-120H560v-80h120v-120h80v120h120v80H760v120h-80Zm-480-80q-33 0-56.5-23.5T120-240v-480q0-33 23.5-56.5T200-800h40v-80h80v80h240v-80h80v80h40q33 0 56.5 23.5T760-720v244q-20-3-40-3t-40 3v-84H200v320h280q0 20 3 40t11 40H200Zm0-480h480v-80H200v80Zm0 0v-80 80Z">
-            </path>
-        </symbol>
-        <symbol id="skiconsBranch" viewBox="0 0 19 17">
-            <path
-                d="M1.16667 2.24999V0.166656H17.8333V2.24999H1.16667ZM1.16667 16.8333V10.5833H0.125V8.49999L1.16667 3.29166H17.8333L18.875 8.49999V10.5833H17.8333V16.8333H15.75V10.5833H11.5833V16.8333H1.16667ZM3.25 14.75H9.5V10.5833H3.25V14.75ZM2.26042 8.49999H16.7396L16.1146 5.37499H2.88542L2.26042 8.49999Z" />
-        </symbol>
-        <symbol id="skIconsArrowForward" viewBox="0 0 24 24">
-            <path d="M0 0h24v24H0V0z" fill="none" />
-            <path d="M12 4l-1.41 1.41L16.17 11H4v2h12.17l-5.58 5.59L12 20l8-8-8-8z" />
-        </symbol>
-        <symbol id="skIconBookAppointment" viewBox="0 0 24 24">
-            <g>
-                <rect fill="none" height="24" width="24" />
-            </g>
-            <g>
-                <path
-                    d="M19,4h-1V2h-2v2H8V2H6v2H5C3.89,4,3.01,4.9,3.01,6L3,20c0,1.1,0.89,2,2,2h14c1.1,0,2-0.9,2-2V6C21,4.9,20.1,4,19,4z M19,20 H5V10h14V20z M19,8H5V6h14V8z M9,14H7v-2h2V14z M13,14h-2v-2h2V14z M17,14h-2v-2h2V14z M9,18H7v-2h2V18z M13,18h-2v-2h2V18z M17,18 h-2v-2h2V18z" />
-            </g>
-        </symbol>
-    </svg>
-</template>
