+++ src/components/template/CoreWebVitals.vue
