+++ src/components/template/WpPhotos.vue
